remove ts-ignore, add SoundConfig
